refactor: implement a singleton approach for the Risk instance provider
