+++ checkout/src/main/java/com/checkout/tokenization/usecase/RiskSdkUseCase.kt
@@ -3,9 +3,6 @@ package com.checkout.tokenization.usecase
 import android.content.Context
 import com.checkout.base.model.Environment
 import com.checkout.base.usecase.UseCase
-import com.checkout.risk.Risk
-import com.checkout.risk.RiskConfig
-import com.checkout.risk.RiskEnvironment
 import com.checkout.tokenization.model.TokenResult
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -15,7 +12,7 @@ internal class RiskSdkUseCase(
     private val environment: Environment,
     private val context: Context,
     private val publicKey: String,
-    private val riskInstanceProvider: RiskInstanceProvider = RiskInstanceProvider(),
+    private val riskInstanceProvider: RiskInstanceProvider,
 ) : UseCase<TokenResult<String>, Unit> {
     override fun execute(data: TokenResult<String>) {
         CoroutineScope(Dispatchers.IO).launch {
@@ -27,28 +24,5 @@ internal class RiskSdkUseCase(
                 is TokenResult.Failure -> {}
             }
         }
-    }
-}
-
-internal class RiskInstanceProvider() {
-    suspend fun provide(
-        context: Context,
-        publicKey: String,
-        environment: Environment,
-    ): Risk? {
-        val riskEnvironment =
-            when (environment) {
-                Environment.PRODUCTION -> RiskEnvironment.PRODUCTION
-                Environment.SANDBOX -> RiskEnvironment.SANDBOX
-            }
-
-        return Risk.getInstance(
-            context,
-            RiskConfig(
-                publicKey = publicKey,
-                environment = riskEnvironment,
-                framesMode = true,
-            ),
-        )
     }
 }
